+++ examples/custom_routes/lib/pwa/worker.dart
@@ -8,6 +8,7 @@ Worker createWorker() {
   DynamicCache youtubeThumbnails =
       new DynamicCache('youtube', maxEntries: 10, skipDiskCache: true);
 
-  worker.router.get('https://i.ytimg.com/vi/', youtubeThumbnails.cacheFirst);
+  worker.router
+      .registerGetUrl('https://i.ytimg.com/vi/', youtubeThumbnails.cacheFirst);
   return worker;
 }